+++ up42/initialization.py
@@ -1,26 +1,13 @@
 import logging
 import warnings
-from typing import List, Optional
+from typing import Optional
 
-from up42 import (
-    asset,
-    catalog,
-    job,
-    jobcollection,
-    jobtask,
-    main,
-    order,
-    project,
-    storage,
-    tasking,
-    utils,
-    webhooks,
-    workflow,
-)
+from up42 import asset, catalog, job, jobtask, main, order, storage, tasking, utils, webhooks, workflow
 
 logger = utils.get_logger(__name__, level=logging.INFO)
 
 DEPRECATION_MESSAGE = "after May 15th, 2024, and will be replaced by new processing functionalities."
+INITIALIZED_MSG = "Initialized %s"
 
 
 def _get_project_id(project_id: Optional[str]) -> str:
@@ -34,26 +21,6 @@ def _get_project_id(project_id: Optional[str]) -> str:
     if not result:
         raise ValueError("Project ID can't be null")
     return result
-
-
-@main.check_auth
-def initialize_project(project_id: Optional[str] = None) -> project.Project:
-    """
-    Returns the correct Project object (has to exist on UP42).
-    Args:
-        project_id: The UP42 project id
-    """
-    warnings.warn(
-        "Projects are getting deprecated. The current analytics platform will be discontinued "
-        f"{DEPRECATION_MESSAGE}",
-        DeprecationWarning,
-    )
-    up42_project = project.Project(
-        auth=main.get_auth_safely(),
-        project_id=_get_project_id(project_id=project_id),
-    )
-    logger.info("Initialized %s", up42_project)
-    return up42_project
 
 
 @main.check_auth
@@ -90,7 +57,7 @@ def initialize_workflow(workflow_id: str, project_id: Optional[str] = None) -> w
         workflow_id=workflow_id,
         project_id=_get_project_id(project_id=project_id),
     )
-    logger.info("Initialized %s", up42_workflow)
+    logger.info(INITIALIZED_MSG, up42_workflow)
     return up42_workflow
 
 
@@ -111,7 +78,7 @@ def initialize_job(job_id: str, project_id: Optional[str] = None) -> job.Job:
         job_id=job_id,
         project_id=_get_project_id(project_id=project_id),
     )
-    logger.info("Initialized %s", up42_job)
+    logger.info(INITIALIZED_MSG, up42_job)
     return up42_job
 
 
@@ -135,38 +102,8 @@ def initialize_jobtask(jobtask_id: str, job_id: str, project_id: Optional[str] =
         job_id=job_id,
         project_id=_get_project_id(project_id=project_id),
     )
-    logger.info("Initialized %s", job_task)
+    logger.info(INITIALIZED_MSG, job_task)
     return job_task
-
-
-@main.check_auth
-def initialize_jobcollection(job_ids: List[str], project_id: Optional[str] = None) -> jobcollection.JobCollection:
-    """
-    Returns a JobCollection object (the referenced jobs have to exist on UP42).
-    Args:
-        job_ids: List of UP42 job_ids
-        project_id: The id of the UP42 project, containing the jobs
-    """
-    warnings.warn(
-        "Job Collections are getting deprecated. The current analytics platform will be discontinued "
-        f"{DEPRECATION_MESSAGE}",
-        DeprecationWarning,
-    )
-    jobs = [
-        job.Job(
-            auth=main.get_auth_safely(),
-            job_id=job_id,
-            project_id=_get_project_id(project_id=project_id),
-        )
-        for job_id in job_ids
-    ]
-    job_collection = jobcollection.JobCollection(
-        auth=main.get_auth_safely(),
-        project_id=_get_project_id(project_id=project_id),
-        jobs=jobs,
-    )
-    logger.info("Initialized %s", job_collection)
-    return job_collection
 
 
 @main.check_auth
@@ -185,7 +122,7 @@ def initialize_order(order_id: str) -> order.Order:
         order_id: The UP42 order_id
     """
     up42_order = order.Order(auth=main.get_auth_safely(), order_id=order_id)
-    logger.info("Initialized %s", up42_order)
+    logger.info(INITIALIZED_MSG, up42_order)
     return up42_order
 
 
@@ -197,7 +134,7 @@ def initialize_asset(asset_id: str) -> asset.Asset:
         asset_id: The UP42 asset_id
     """
     up42_asset = asset.Asset(auth=main.get_auth_safely(), asset_id=asset_id)
-    logger.info("Initialized %s", up42_asset)
+    logger.info(INITIALIZED_MSG, up42_asset)
     return up42_asset
 
 
@@ -209,5 +146,5 @@ def initialize_webhook(webhook_id: str) -> webhooks.Webhook:
         webhook_id: The UP42 webhook_id
     """
     webhook = webhooks.Webhook(auth=main.get_auth_safely(), webhook_id=webhook_id)
-    logger.info("Initialized %s", webhook)
+    logger.info(INITIALIZED_MSG, webhook)
     return webhook