+++ CHANGELOG.md
@@ -29,6 +29,16 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 1.0.0a2
+
+**Apr 15, 2024**
+
+- Dropped deprecated JobCollection functions - info, status, apply, download_results
+- Dropped deprecated Project functions - info, get_workflows, get_project_settings, get_jobs
+- Dropped Project's fixtures and tests
+- Dropped JobCollection's fixtures and tests
+- Dropped Workflow's get_jobs function
+
 ## 1.0.0a1
 
 **Apr 9, 2024**